requirements.txt
anthropic>=0.39.0
openai>=1.54.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
GitPython>=3.1.40
pygments>=2.17.2
aiohttp>=3.9.1
pydantic>=2.5.3
tenacity>=8.2.3

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
